Add files via upload
For the program to function you will need Python 3 and Tkinter installed.

You will need both files to make it function correctly.

1.  ‘ConundrumGUI3New.pyw’

This is the main body of the program. I have tried to put some comments in for people who want to understand the functions more easily, but as I wrote it several years ago, much of it is confusing even to me.

2. ‘newdict.txt’

This is the dictionary file. It has been prepared in such a way that it will be imported directly as a list, I\so it is formatted in this way i.e: [‘apple’, ‘amber’, ‘avo’]

Just drop both these into the same folder and you’ll be able to use it directly/

The program is fully GUI oriented and I hope it will shed some light on how to build useful software with Tkinter and Python.